Fixes search pagination bugs
